# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 漏洞管理平台 API
  version: 1.0.0
  description: |
    漏洞管理平台对外接口。包含两类调用方：

    - **扫描工具**（机器）：使用 `X-API-Key`，调用 `/ingest/**` 上报漏洞与正负样本。
    - **平台前端**（浏览器）：使用 `Authorization: Bearer <JWT>`，调用查询与处置接口。

    统一响应格式：`{ code, message, data, traceId }`，`code = 0` 表示业务成功。
    时间统一为 ISO 8601 UTC，字段统一小驼峰。
    兼容策略：只增字段，不删不改语义。
  contact:
    name: 平台侧
servers:
  - url: http://localhost:4010/api/v1
    description: 本地 Mock（prism mock openapi.yaml）
  - url: http://localhost:3000/api/v1
    description: 本地开发
  - url: https://vuln.example.com/api/v1
    description: 生产
...
security:
  - bearerAuth: []
paths:
...
  /stats/overview:
    get:
      tags: [Stats]
      summary: 看板总览数据
      parameters:
        - name: projectId
          in: query
          schema: { type: integer }
        - name: days
          in: query
          schema: { type: integer, default: 30, minimum: 1, maximum: 365 }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/StatsOverview' }
  /stats/trend:
    get:
      tags: [Stats]
      summary: 漏洞趋势（按天）
      parameters:
        - name: days
          in: query
          schema: { type: integer, default: 30 }
        - name: projectId
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          list:
                            type: array
                            items:
                              type: object
                              properties:
                                date: { type: string, format: date }
                                newCount: { type: integer }
                                fixedCount: { type: integer }
                                openCount: { type: integer }
  /stats/severity:
    get:
      tags: [Stats]
      summary: 漏洞等级分布
      parameters:
        - name: projectId
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          list:
                            type: array
                            items:
                              type: object
                              properties:
                                severity: { $ref: '#/components/schemas/Severity' }
                                count: { type: integer }
                                openCount: { type: integer }
                                fixedCount: { type: integer }
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: 扫描工具使用的长期凭证，形如 `vuln_sk_xxxxxxxx`。仅授予 `ingest` 权限范围。
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 平台前端登录后获得，有效期 8 小时。
...
  schemas:
    ApiResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: 0 表示业务成功，非 0 见错误码表
        message: { type: string }
        data: { nullable: true }
        traceId: { type: string }
...
    Severity:
      type: string
      enum: [critical, high, medium, low, info]
      description: 严重 / 高危 / 中危 / 低危 / 提示
...
    StatsOverview:
      type: object
      properties:
        vulnTotal: { type: integer, description: 漏洞总数（不含误报与已忽略） }
        vulnOpen: { type: integer, description: open + confirmed + fixing }
        vulnCritical: { type: integer }
        vulnHigh: { type: integer }
        newInPeriod: { type: integer }
        fixedInPeriod: { type: integer }
        projectCount: { type: integer }
        scanCount: { type: integer }
        sampleCount: { type: integer }
        avgFixHours: { type: number, description: 平均修复时长（小时） }